Replace form field generation classes
